requirements.txt
fastapi>=0.89.1
requests
uvicorn
dvclive==1.2.0
dvc[all]
GitPython==3.1.29
python-dotenv>=0.21.0